requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==2.3.7
